purchased items screen and check internet connection

+++ app/src/main/java/com/example/ecommerceapp/presentation/purchased_item/PurchasedItemsScreen.kt
@@ -79,7 +79,7 @@ fun PurchasedItemsScreen(
                 scope.launch {
                     snackbarHostState.showSnackbar(
                         message = error.message ?: "Unknown Error",
-                        duration = SnackbarDuration.Long
+                        duration = SnackbarDuration.Short
                     )
                 }
             }
